+++ typeit/schema/__init__.py
@@ -1,24 +1,17 @@
 import enum as std_enum
-from typing import Type, NamedTuple, Sequence, Union, Any
+import pathlib
+from typing import Type, NamedTuple, Sequence, Union, Any, Mapping, Tuple, Set
 
 import colander as col
 
-from .definitions import OverridesT
-from .sums import SumType
+from ..definitions import OverridesT
+from ..sums import SumType
+from .. import interface as iface
+
+from . import primitives
 
 
 EnumLike = Union[std_enum.Enum, SumType]
-
-
-class AcceptEverything(col.SchemaType):
-    """ A schema type to correspond to typing.Any, i.e. allows
-    any data to pass through the type constructor.
-    """
-    def serialize(self, node, appstruct):
-        return appstruct
-
-    def deserialize(self, node, cstruct):
-        return cstruct
 
 
 class Enum(col.Str):
@@ -27,9 +20,6 @@ class Enum(col.Str):
         self.typ = typ
 
     def serialize(self, node, appstruct):
-        """ Default colander integer serializer returns a string representation
-        of a number, whereas we want identical representation of the original data.
-        """
         if appstruct is col.null:
             return appstruct
         r = super().serialize(node, appstruct.value)
@@ -45,10 +35,32 @@ class Enum(col.Str):
             raise col.Invalid(node, f'Invalid variant of {self.typ.__name__}', cstruct)
 
 
-class Structure(col.Mapping):
+class Path(col.Str):
+    def __init__(self, typ: Type[pathlib.PurePath], *args, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
+        self.typ = typ
 
+    def serialize(self, node, appstruct: Union[col._null, pathlib.PurePath]):
+        if appstruct is col.null:
+            return appstruct
+        r = super().serialize(node, str(appstruct))
+        return r
+
+    def deserialize(self, node, cstruct) -> pathlib.PurePath:
+        r = super().deserialize(node, cstruct)
+        if r is col.null:
+            return r
+        try:
+            return self.typ(r)
+        except TypeError:
+            raise col.Invalid(node, f'Invalid variant of {self.typ.__name__}', cstruct)
+
+
+class Structure(col.Mapping):
+    """ SchemaNode for NamedTuples and derived types.
+    """
     def __init__(self,
-                 typ: Type[NamedTuple],
+                 typ: Type[iface.IType],
                  overrides: OverridesT,
                  unknown: str = 'ignore') -> None:
         super().__init__(unknown)
@@ -89,11 +101,18 @@ class Structure(col.Mapping):
 
 
 class UnionNode(col.SchemaType):
-    def __init__(self,
-                 variants: Sequence[col.SchemaNode]) -> None:
+    def __init__(
+        self,
+        variants: Sequence[Union[col.SchemaNode, col.SequenceSchema, col.TupleSchema]],
+        primitives_registry: Union[
+            Mapping[Type, primitives.PrimitiveSchemaTypeT],
+            Mapping[Type, primitives.NonStrictPrimitiveSchemaTypeT],
+        ]
+    ) -> None:
         super().__init__()
+        self.primitives_registry = primitives_registry
         self.variants = variants
-        self.variant_types = {x.typ for x in variants}
+        self.variant_types: Set[col.SchemaNode] = {x.typ for x in variants}
 
     def deserialize(self, node, cstruct):
         if cstruct is None:
@@ -113,7 +132,7 @@ class UnionNode(col.SchemaType):
         # fact that both str() and int() constructors can happily
         # handle that value and return one of the expected variants
         # (but incorrectly!)
-        prim_schema_type = BUILTIN_TO_SCHEMA_TYPE.get(type(cstruct))
+        prim_schema_type = self.primitives_registry.get(type(cstruct))
         if prim_schema_type in self.variant_types:
             try:
                 return prim_schema_type.deserialize(node, cstruct)
@@ -138,13 +157,13 @@ class UnionNode(col.SchemaType):
             cstruct
         )
 
-    def serialize(self, node, appstruct: NamedTuple):
+    def serialize(self, node, appstruct: Any):
         if appstruct in (col.null, None):
             return None
 
         struct_type = type(appstruct)
 
-        prim_schema_type = BUILTIN_TO_SCHEMA_TYPE.get(struct_type)
+        prim_schema_type = self.primitives_registry.get(struct_type)
         if prim_schema_type in self.variant_types:
             try:
                 return prim_schema_type.serialize(node, appstruct)
@@ -157,45 +176,41 @@ class UnionNode(col.SchemaType):
         )
 
         for variant in remaining_variants:
-            if variant.typ.typ is not struct_type:
-                continue
-            return variant.serialize(appstruct)
+            # Sequences and tuples require special treatment here:
+            # since there is no direct reference to the target python data type
+            # through variant.typ.typ that we could use to compare this variant
+            # with appstruct's type, we just check if the appstruct is a list-like
+            # object. And if it is, we apply SequenceSchema's serializer on it,
+            # otherwise we skip this variant (we need to do that to make sure that
+            # a Union variant matches appstruct's type as close as possible)
+            # if isinstance(struct_type, (list, tuple, set))
+            if isinstance(variant, col.SequenceSchema):
+                if not isinstance(appstruct, list):
+                    continue
+                try:
+                    return variant.serialize(appstruct)
+                except col.Invalid:
+                    continue
 
+            elif isinstance(variant, col.TupleSchema):
+                if not isinstance(appstruct, tuple):
+                    continue
+                try:
+                    return variant.serialize(appstruct)
+                except col.Invalid:
+                    continue
 
-class Int(col.Int):
+            else:
+                # schema.SchemaNode
+                if variant.typ.typ is not struct_type:
+                    continue
+                return variant.serialize(appstruct)
 
-    def serialize(self, node, appstruct):
-        """ Default colander integer serializer returns a string representation
-        of a number, whereas we want identical representation of the original data.
-        """
-        r = super().serialize(node, appstruct)
-        if r is col.null:
-            return r
-        return int(r)
-
-
-class Bool(col.Bool):
-
-    def serialize(self, node, appstruct):
-        """ Default colander bool serializer returns a string representation
-        of a boolean flag, whereas we want identical representation of the original data.
-        """
-        r = super().serialize(node, appstruct)
-        if r is col.null:
-            return r
-        return {'false': False, 'true': True}[r]
-
-
-class Str(col.Str):
-
-    def serialize(self, node, appstruct):
-        """ Default colander str serializer serializes None as 'None',
-        whereas we want identical representation of the original data.
-        """
-        r = super().serialize(node, appstruct)
-        if r in (col.null, 'None'):
-            return None
-        return r
+        raise col.Invalid(
+            node,
+            'None of the variants matches provided structure. ',
+            appstruct
+        )
 
 
 class SetSchema(col.SequenceSchema):
@@ -220,10 +235,30 @@ class SchemaNode(col.SchemaNode):
         return f'SchemaNode({self.typ})'
 
 
-BUILTIN_TO_SCHEMA_TYPE = {
-    Any: AcceptEverything(),
-    str: Str(allow_empty=True),
-    int: Int(),
-    float: col.Float(),
-    bool: Bool(),
+# Maps primitive types that appear in type signatures
+# to colander SchemaNodes responsible for serialization/deserialization
+BUILTIN_TO_SCHEMA_TYPE: Mapping[Type, primitives.PrimitiveSchemaTypeT] = {
+    Any: primitives.AcceptEverything(),
+    str: primitives.Str(allow_empty=True),
+    int: primitives.Int(),
+    float: primitives.Float(),
+    bool: primitives.Bool(),
+}
+
+
+NON_STRICT_BUILTIN_TO_SCHEMA_TYPE: Mapping[Type, primitives.NonStrictPrimitiveSchemaTypeT] = {
+    Any: primitives.AcceptEverything(),
+    str: primitives.NonStrictStr(allow_empty=True),
+    int: primitives.NonStrictInt(),
+    float: primitives.NonStrictFloat(),
+    bool: primitives.NonStrictBool(),
+}
+
+
+_SUBCLASS_BASED_TO_SCHEMA_NODE: Mapping[
+    Tuple[Type, ...], Type[col.SchemaNode],
+] = {
+    (std_enum.Enum, SumType): Enum,
+    # Pathlib's PurePath and its derivatives
+    (pathlib.PurePath,): Path,
 }