+++ CHANGELOG.rst
@@ -2,6 +2,15 @@
 CHANGELOG
 =========
 
+0.10.0
+============
+
+* Fix bug in parsing union types with sequence variants.
+* Primitive types switched to strict matching.
+* Non-strict primitives flag `NON_STRICT_PRIMITIVES` is available for overrides.
+* Added support for `typing.Mapping`
+* Added support for `pathlib.Path`
+
 0.9.0
 ============